refactor: presentation common 만들기
