+++ android/app/src/main/java/com/woowacourse/ody/presentation/common/BackListener.kt
@@ -1,0 +1,5 @@
+package com.woowacourse.ody.presentation.common
+
+interface BackListener {
+    fun onBack()
+}
