+++ android/app/src/main/java/com/woowacourse/ody/presentation/common/NextListener.kt
@@ -1,0 +1,5 @@
+package com.woowacourse.ody.presentation.common
+
+interface NextListener {
+    fun onNext()
+}
