+++ android/app/src/main/java/com/woowacourse/ody/presentation/common/ViewPagerAdapter.kt
@@ -1,0 +1,14 @@
+package com.woowacourse.ody.presentation.common
+
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
+import androidx.viewpager2.adapter.FragmentStateAdapter
+
+class ViewPagerAdapter(
+    fragmentActivity: FragmentActivity,
+    private val fragments: List<Fragment>,
+) : FragmentStateAdapter(fragmentActivity) {
+    override fun getItemCount(): Int = fragments.size
+
+    override fun createFragment(position: Int): Fragment = fragments[position]
+}
