+++ android/app/src/main/java/com/woowacourse/ody/presentation/intro/IntroActivity.kt
@@ -16,7 +16,7 @@ import com.woowacourse.ody.R
 import com.woowacourse.ody.databinding.ActivityIntroBinding
 import com.woowacourse.ody.presentation.invitecode.InviteCodeActivity
 import com.woowacourse.ody.presentation.meetinginfo.MeetingInfoActivity
-import com.woowacourse.ody.util.observeEvent
+import com.woowacourse.ody.presentation.common.observeEvent
 
 class IntroActivity : AppCompatActivity() {
     private val vm: IntroViewModel by viewModels()
