+++ android/app/src/main/java/com/woowacourse/ody/presentation/invitecode/InviteCodeActivity.kt
@@ -11,8 +11,8 @@ import com.woowacourse.ody.R
 import com.woowacourse.ody.data.remote.ody.repository.DefaultMeetingRepository
 import com.woowacourse.ody.databinding.ActivityInviteCodeBinding
 import com.woowacourse.ody.presentation.joininfo.JoinInfoActivity
-import com.woowacourse.ody.presentation.meetinginfo.BackListener
-import com.woowacourse.ody.util.observeEvent
+import com.woowacourse.ody.presentation.common.BackListener
+import com.woowacourse.ody.presentation.common.observeEvent
 
 class InviteCodeActivity : AppCompatActivity(), BackListener {
     private val binding: ActivityInviteCodeBinding by lazy {
