--- android/app/src/main/java/com/woowacourse/ody/presentation/meetinginfo/BackListener.kt
@@ -1,5 +1,0 @@
-package com.woowacourse.ody.presentation.meetinginfo
-
-interface BackListener {
-    fun onBack()
-}
