+++ android/app/src/main/java/com/woowacourse/ody/presentation/meetinginfo/MeetingInfoViewModel.kt
@@ -7,8 +7,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.map
 import com.woowacourse.ody.domain.model.GeoLocation
 import com.woowacourse.ody.presentation.address.AddressValidator
-import com.woowacourse.ody.util.Event
-import com.woowacourse.ody.util.emit
+import com.woowacourse.ody.presentation.common.Event
+import com.woowacourse.ody.presentation.common.emit
 import java.time.LocalDateTime
 import java.time.LocalTime
