--- android/app/src/main/java/com/woowacourse/ody/util/NextListener.kt
@@ -1,5 +1,0 @@
-package com.woowacourse.ody.util
-
-interface NextListener {
-    fun onNext()
-}
