+++ android/app/src/main/res/layout/activity_invite_code.xml
@@ -10,7 +10,7 @@
 
         <variable
             name="backListener"
-            type="com.woowacourse.ody.presentation.meetinginfo.BackListener" />
+            type="com.woowacourse.ody.presentation.common.BackListener" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
