+++ android/app/src/main/res/layout/activity_meeting_info.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="backListener"
-            type="com.woowacourse.ody.presentation.meetinginfo.BackListener" />
+            type="com.woowacourse.ody.presentation.common.BackListener" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
