+++ android/app/src/main/res/layout/toolbar_back.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="backListener"
-            type="com.woowacourse.ody.presentation.meetinginfo.BackListener" />
+            type="com.woowacourse.ody.presentation.common.BackListener" />
     </data>
 
     <androidx.appcompat.widget.Toolbar
